[INTERPRETER] * BUG: Synchronize the 'gambas.h' header with the internal date management   changes of rev. #5871.
git-svn-id: svn://localhost/gambas/trunk@5873 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ main/share/gambas.h
@@ -643,14 +643,14 @@ typedef
 
 typedef
 	struct {
-		short year;
-		short month;
-		short day;
-		short hour;
-		short min;
-		short sec;
-		short weekday;
-		short msec;
+		int year;
+		int month;
+		int day;
+		int hour;
+		int min;
+		int sec;
+		int weekday;
+		int msec;
 		}
 	GB_DATE_SERIAL;
